fix: render file upload surface inline instead of fallback chip
The file_upload surface was classified as a "chip only" surface in
InlineSurfaceRouter, causing it to render as a tiny fallback chip instead
of the full drag-and-drop FileUploadSurfaceView. Moved the view to the
shared module and wired it up in the inline router.
